Adjust component init
Changed components to be created on weapon creation. There's now a separate step added for when the Owner is applied to the archetype, upon which the AutoAim property is removed or kept (if applicable).

Fixes a bug with Ammo not being set on drop since Owner is not set at that point.