v3.5.0 - Shrapnel, HealthShotMod, ChargeExponent
Added a field to HealthMod to apply to the player hit (if hitting a player).
Fixed Noise doing nothing with UseNoiseSystem if FakeAlertRadius was less than or equal to WakeUpRadius.
Fixed FireShot doing bullet FX from the gun when firing from a hit position.
Changed EntryPoint directly calling functions to using invoke attributes.
Attempted a fix for clients considering enemies as foamed.